requirements.txt
flet
flet-fastapi
supabase
python-dotenv
